convert double to Int

+++ app/src/main/java/com/example/vladimirdvortsov/weatherkotlin/ui/MainActivity.kt
@@ -78,7 +78,7 @@ class MainActivity : AppCompatActivity() {
             if (!temp.text.isNullOrEmpty()){
                 if (isCelsia) {
                     metric_type.text = " °F"
-                    temp.text = (temp.text.toString().toDouble() - 273).toString()
+                    temp.text = (temp.text.toString().toInt() - 273).toString()
                     isCelsia = false
                     //celsia.setBackgroundColor(ContextCompat.getColor(context, R.color.colorGrey))
                     //fahrenheit.setBackgroundColor(ContextCompat.getColor(context, R.color.cardBackground))
@@ -87,7 +87,7 @@ class MainActivity : AppCompatActivity() {
                 }
                 else{
                     metric_type.text = " °C"
-                    temp.text = (temp.text.toString().toDouble() + 273).toString()
+                    temp.text = (temp.text.toString().toInt() + 273).toString()
                     isCelsia = true
 //                    celsia.setBackgroundColor(ContextCompat.getColor(context, R.color.cardBackground))
 //                    fahrenheit.setBackgroundColor(ContextCompat.getColor(context, R.color.colorGrey))
@@ -101,7 +101,7 @@ class MainActivity : AppCompatActivity() {
         fahrenheit.setOnClickListener {
             if (!temp.text.isNullOrEmpty()){
                 if (isCelsia) {
-                    temp.text = (temp.text.toString().toDouble() - 273).toString()
+                    temp.text = (temp.text.toString().toInt() - 273).toString()
                     metric_type.text = " °F"
                     isCelsia = false
 //                    celsia.setBackgroundColor(ContextCompat.getColor(context, R.color.colorGrey))
@@ -110,7 +110,7 @@ class MainActivity : AppCompatActivity() {
                     fahrenheit.resources.getDrawable(R.drawable.right_white)
                 }
                 else{
-                    temp.text = (temp.text.toString().toDouble() + 273).toString()
+                    temp.text = (temp.text.toString().toInt() + 273).toString()
                     metric_type.text = " °C"
                     isCelsia = true
 //                    celsia.setBackgroundColor(ContextCompat.getColor(context, R.color.cardBackground))
